feat(bootstrap): add addons to select
fix #1380
