+++ src/bootstrap/addons/src/addon.extension.ts
