+++ src/bootstrap/addons/src/public_api.ts
@@ -1,0 +1,1 @@
+export { FormlyBootstrapAddonsModule } from './addons.module';
