+++ src/bootstrap/ng-package.json
@@ -4,6 +4,7 @@
   "deleteDestPath": false,
   "lib": {
     "entryFile": "src/public_api.ts",
+    "addons": [{ "path": "addons" }],
     "externals": {
       "@ngx-formly/core": "@ngx-formly/core"
     }
