--- src/bootstrap/src/lib/run/addon.ts
@@ -1,13 +1,0 @@
-import { FormlyConfig, FormlyFieldConfig } from '@ngx-formly/core';
-
-export class TemplateAddons {
-  run(fc: FormlyConfig) {
-    fc.templateManipulators.postWrapper.push((field: FormlyFieldConfig) => {
-      if (field && field.templateOptions && (field.templateOptions.addonLeft || field.templateOptions.addonRight)) {
-        return 'addons';
-      }
-
-      return '';
-    });
-  }
-}
