+++ src/bootstrap/src/lib/types/input.ts
@@ -9,10 +9,6 @@ import { FieldType } from '@ngx-formly/core';
       <input type="number" [formControl]="formControl" class="form-control" [formlyAttributes]="field" [class.is-invalid]="showError">
     </ng-template>
   `,
-  host: {
-    '[class.d-inline-flex]': 'to.addonLeft || to.addonRight',
-    '[class.custom-file]': 'to.addonLeft || to.addonRight',
-  },
 })
 export class FormlyFieldInput extends FieldType {
   get type() {
