+++ src/bootstrap/src/lib/wrappers/wrappers.ts
@@ -1,2 +1,1 @@
 export { FormlyWrapperFormField } from './form-field.wrapper';
-export { FormlyWrapperAddons } from './addons';
